requirements.txt
# for working with IP address
ipaddress>=1.0.17
# https://github.com/kelproject/pykube/issues/29#issuecomment-230026930
backports.ssl_match_hostname

requests[security]>=2.10.0
boto==2.39.0
boto3>=1.3.1
botocore>=1.3.26
click>=6.2

# pykube>=0.12.0 we need master branch
git+git://github.com/kelproject/pykube.git@8a5cd8f6babc013e11f02b9594fb534e9c58689f

# for tests
nose>=1.3.7
PyYAML>=3.11
moto>=0.4.25
mock>=2.0.0
